for deployment

+++ analysis/analysis_subsample2.py
@@ -141,11 +141,11 @@ def analyis_subsample(id, config, soa):
             evokeds_h2 = {cond: [] for cond in [cond1, cond2]}
             
 
-            subsample_epochs1_h1 = epochs1[idx1[:num]][:(num/2)]
-            subsample_epochs1_h2 = epochs1[idx1[:num]][(num/2):]
+            subsample_epochs1_h1 = epochs1[idx1[:num]][:int(num/2)]
+            subsample_epochs1_h2 = epochs1[idx1[:num]][int(num/2):]
 
-            subsample_epochs2_h1 = epochs2[idx2[:num]][:(num/2)]
-            subsample_epochs2_h2 = epochs2[idx2[:num]][(num/2):]
+            subsample_epochs2_h1 = epochs2[idx2[:num]][:int(num/2)]
+            subsample_epochs2_h2 = epochs2[idx2[:num]][int(num/2):]
 
             # average
             evokeds_h1[cond1].append(subsample_epochs1_h1.average())
